src/requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
djangorestframework==3.11.0
get==2019.4.13
idna==2.9
lxml==4.5.0
post==2019.4.13
public==2019.4.13
python-dotenv==0.12.0
pytz==2019.3
query-string==2019.4.13
request==2019.4.13
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.8
